requirements.txt
certifi==2021.5.30
cffi==1.14.5
cryptography==3.4.7
pycparser==2.20
pycryptodome==3.10.1
pyDes==2.0.1
wincertstore==0.2
